requirements.txt
aiohttp==3.12.1
aiogram==3.22.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
asyncpg==0.31.0
pycryptodome==3.23.0
aiofiles==24.1.0
python-dateutil==2.9.0
Pillow==11.1.0